+++ EleFitsData/EleFitsData/impl/Column.hpp
@@ -29,14 +29,18 @@ template <typename T, long N, typename TContainer>
 ColumnContainer<T, N, TContainer>::ColumnContainer() : ColumnContainer(ColumnContainer<T, N, TContainer>::Info {}) {}
 
 template <typename T, long N, typename TContainer>
-ColumnContainer<T, N, TContainer>::ColumnContainer(ColumnContainer<T, N, TContainer>::Info info, long size) :
-    ColumnContainer<T, N, TContainer>::Base(size), m_info(std::move(info)) {
-  // FIXME assert size() % m_info.elementCountPerEntry() == 0;
+ColumnContainer<T, N, TContainer>::ColumnContainer(ColumnContainer<T, N, TContainer>::Info info, long rowCount) :
+    ColumnContainer<T, N, TContainer>::Base(info.elementCount() * rowCount), m_info(std::move(info)) {
+  // FIXME assert size() % m_info.elementCount() == 0;
 }
 
 template <typename T, long N, typename TContainer>
-ColumnContainer<T, N, TContainer>::ColumnContainer(ColumnContainer<T, N, TContainer>::Info info, long size, T* data) :
-    ColumnContainer<T, N, TContainer>::Base(data, data + size), m_info(std::move(info)) {
+ColumnContainer<T, N, TContainer>::ColumnContainer(
+    ColumnContainer<T, N, TContainer>::Info info,
+    long rowCount,
+    T* data) :
+    ColumnContainer<T, N, TContainer>::Base(data, data + info.elementCount() * rowCount),
+    m_info(std::move(info)) {
   // FIXME idem
 }
 
@@ -59,8 +63,9 @@ void ColumnContainer<T, N, TContainer>::rename(const std::string& name) {
 
 template <typename T, long N, typename TContainer>
 void ColumnContainer<T, N, TContainer>::reshape(long repeatCount) {
-  // FIXME check that elementCount() % repeatCount() = 0, but for strings!
-  m_info.shape[0] = repeatCount;
+  // FIXME check that elementCount() % repeatCount = 0, but for strings!
+  // FIXME check that shape is valid
+  m_info.repeatCount = repeatCount;
 }
 
 template <typename T, long N, typename TContainer>
@@ -70,12 +75,12 @@ long ColumnContainer<T, N, TContainer>::elementCount() const {
 
 template <typename T, long N, typename TContainer>
 long ColumnContainer<T, N, TContainer>::rowCount() const {
-  return elementCount() / m_info.elementCountPerEntry();
+  return elementCount() / m_info.elementCount();
 }
 
 template <typename T, long N, typename TContainer>
 const T& ColumnContainer<T, N, TContainer>::operator()(long row, long repeat) const {
-  const long index = row * m_info.elementCountPerEntry() + repeat;
+  const long index = row * m_info.elementCount() + repeat;
   return *(this->data() + index);
 }
 
@@ -87,7 +92,7 @@ T& ColumnContainer<T, N, TContainer>::operator()(long row, long repeat) {
 template <typename T, long N, typename TContainer>
 const T& ColumnContainer<T, N, TContainer>::at(long row, long repeat) const {
   OutOfBoundsError::mayThrow("Cannot access row index", row, {-rowCount(), rowCount() - 1});
-  const auto bound = m_info.elementCountPerEntry();
+  const auto bound = m_info.elementCount();
   OutOfBoundsError::mayThrow("Cannot access repeat index", repeat, {-bound, bound - 1});
   const long boundedRow = row < 0 ? rowCount() + row : row;
   const long boundedRepeat = repeat < 0 ? bound + repeat : repeat;