+++ EleFitsData/EleFitsData/impl/TestColumn.hpp
@@ -68,15 +68,15 @@ RandomScalarColumn<std::string>::RandomScalarColumn(long size, std::string min, 
     VecColumn<std::string>({"SCALAR", "m", 1}, generateRandomVector<std::string>(size, min, max)) {
   for (const auto& v : vector()) {
     long currentSize = static_cast<long>(v.length() + 1); // +1 for '\0'
-    if (currentSize > info().repeatCount()) {
+    if (currentSize > info().repeatCount) {
       reshape(currentSize);
     }
   }
 }
 
 template <typename T>
-RandomVectorColumn<T>::RandomVectorColumn(long repeatCount, long size, T min, T max) :
-    VecColumn<T>({"VECTOR", "m", repeatCount}, generateRandomVector<T>(repeatCount * size, min, max)) {}
+RandomVectorColumn<T>::RandomVectorColumn(long repeatCount, long rowCount, T min, T max) :
+    VecColumn<T>({"VECTOR", "m", repeatCount}, generateRandomVector<T>(repeatCount * rowCount, min, max)) {}
 
 } // namespace Test
 } // namespace Fits